% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeightPercentiles.r
\name{calcHeightPercentiles}
\alias{calcHeightPercentiles}
\title{Calculate Height Percentiles}
\usage{
calcHeightPercentiles(x, resolution = 30, percentiles = c(.10, .20, .30, .40, .50, .60, .70, .80, .90, .98), pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{percentiles}{a list of numbers specifying the percentiles to be calculated.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
returns a raster stack where each layer is a percentile.
}
\description{
This function calculates percentiles of lidar points.
}
\examples{

}
\author{
Nicholas Kruskamp
}
