% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHeightPointCounts.r
\name{calcHeightPointCounts}
\alias{calcHeightPointCounts}
\title{Count Lidar Points in Height Layers}
\usage{
calcHeightPointCounts(x, resolution = 30, binHeight = NA, binCount = 3, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}

\item{BinHeight}{value to set a fixed height for count layer. If not set, binCount is used.}

\item{BinCount}{value to evenly divide the canopy into count layers. Default is 3 layers.}
}
\value{
a raster stack where each layer is represents the point counts at that height bin.
}
\description{
This function counts the number of points above ground level from lidar data. It will divide the canopy into even height bins, or by number of bins.
}
\examples{

}
\author{
Nicholas Kruskamp
}
