% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyCover.r
\name{canopyCover}
\alias{canopyCover}
\title{Calculate Canopy Cover}
\usage{
canopyCover(x, resolution = 30, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
a raster layer
}
\description{
This function calculates the canopy cover of an area from lidar data. The canopy cover is calculated as the percentage of vegetation first returns to all first returns.
}
\examples{

canopyCoverRaster <- canopyCover(lasdata)

}
\author{
Nicholas Kruskamp
}
