% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopyDensity.r
\name{canopyDensity}
\alias{canopyDensity}
\title{Calculate Canopy Density}
\usage{
canopyDensity(x, resolution = 30, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
a raster layer
}
\description{
This function calculates the canopy density of an area from lidar data. The canopy cover is calculated as the percentage of vegetation returns to all returns.
}
\examples{

canopyDensityRaster <- canopyDensity(lasdata)

}
\author{
Nicholas Kruskamp
}
