% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chm.R
\name{canopyHeightModel}
\alias{canopyHeightModel}
\title{Calculate a canopy height model}
\usage{
canopyHeightModel(x, resolution = 30, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
returns a raster of the canopy height model
}
\description{
This function calculates a canopy height model as the difference between the first return surface and a bare earth dem.
}
\examples{

}
\author{
Nicholas Kruskamp
}
