% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyNoise.r
\name{classifyNoise}
\alias{classifyNoise}
\title{Flags noise points in lidar data.}
\usage{
classifyNoise(x, threshold = 3, nn=7)
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{threshold}{Number specifying the euclidian distance from a given point in which to identify neighborss}

\item{pointClasses}{Number of neighbors a point should have to not be considered an outlier/noise.}
}
\value{
in-place processing of the Spatial Points Data Frame, classifying noise as Classification 7.
}
\description{
This function identifies potential outlier/noise points using a k-nearest neighbor approach, where if a given point does not have "nn" neighbors within "threshold" distance, then it is an outlier/noise.
}
\examples{

}
\author{
Nicholas Kruskamp
}
