% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyVegByHeight.r
\name{classifyVegByHeight}
\alias{classifyVegByHeight}
\title{Classify lidar points by height above ground level.}
\usage{
classifyVegByHeight(x, binHeight = NA, binCount = 3)
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{BinHeight}{value to set a fixed height for classification. If not set, binCount is used.}

\item{BinCount}{value to evenly divide the canopy into classes. Default is 3 layers/classes.}
}
\value{
returns the spatial points data frame with appended classification column.
}
\description{
This function will take normalized Z elevation values and reassign the classification.
}
\examples{

}
\author{
Nicholas Kruskamp
}
