% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDEM.R
\name{createDEM}
\alias{createDEM}
\title{create bare eath digital elevation model (DEM)}
\usage{
createDSM (SpatialPointsDataFrame, res = 10)
}
\arguments{
\item{A}{SpatialPointsDataFrame containing lidar data, at minimum X, Y, and Z values.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}
}
\value{
Returns a raster of the dem elevation values.
}
\description{
This function reads and returns values associated with the LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{

}
\author{
Nicholas Kruskamp
}
