% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDSM.R
\name{createDSM}
\alias{createDSM}
\title{create first return digital surface model (DSM)}
\usage{
createDSM (SpatialPointsDataFrame, res = 10)
}
\arguments{
\item{A}{SpatialPointsDataFrame containing lidar data, at minimum X, Y, and Z values.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}
}
\value{
Returns a raster of the dsm elevation values.
}
\description{
This function performs krige interpolation on all lidar first return to produce a digital surface model.
}
\examples{

}
\author{
Nicholas Kruskamp
}
