% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeByDEM.R
\name{normalizeByDEM}
\alias{normalizeByDEM}
\title{normalize values by height above group from a digital elevation model.}
\usage{
normalizeByDEM(x, y)
}
\arguments{
\item{A}{SpatialPointsDataFrame containing lidar data, at minimum X, Y, and Z values.}

\item{A}{raster dem.}
}
\value{
Returns the SpatialPointsDataFrame with an additional vector of normalized Z values.
}
\description{
This function finds the difference between a non-ground point Z elevation and the coincident cell value Z in X Y space.
}
\examples{

}
\author{
Nicholas Kruskamp
}
