% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeHAG.r
\name{normalizeHAG}
\alias{normalizeHAG}
\title{Normalize Z Values}
\usage{
normalizeHAG (lasdata)
}
\arguments{
\item{lasdata, }{a Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}
}
\value{
returns the spatial points data frame with additional column "Z above ground level (Z_agl)" value.
}
\description{
This function returns the height above ground level for each non ground point. The function finds the nearest ground classified point in X,Y space, and computes the Z different for each non ground point.
}
\examples{

# dem surface raster is interpolated using mean of ground classified values, with default grid cell of 30 units.
lasdata <- normalize(lasdata)

}
\author{
Nicholas Kruskamp
}
