% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointStatistics.r
\name{pointStatistics}
\alias{pointStatistics}
\title{Calculate basic statistics from lidar point data}
\usage{
pointStatistics(x, resolution = 30, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
returns a raster stack where each layer is a metric.
}
\description{
This function calculates statistics from lidar point data, including: maximum, mean, standard deviation, skewness, kurtosis, and quadratic mean.
}
\examples{

}
\author{
Nicholas Kruskamp
}
