% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLASdata_bin.R
\name{readLASdata_bin}
\alias{readLASdata_bin}
\title{Read LiDAR point data}
\usage{
readLASdata_bin (LASfile, short=TRUE)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}

\item{short}{Logical, if TRUE it will return only a 5-column matrix with information on the returned point x, y, z locations, point intensity and the number of return within an individual discrete-return system laser pulse.}

\item{String}{contains the exact proj4string of the input data to project the resulting spatial points data frame.}
}
\value{
Returns a Spatial Points class of the point information stored in the LAS file.
}
\description{
This function reads and returns values associated with the LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{

}
\author{
Original Script: rLiDAR package by Michael Sumner and Carlos Alberto Silva. https://cran.r-project.org/package=rLiDAR Modified by Nicholas Kruskamp to meet (most of) LAS 1.4 spec and output a Spatial Points class.
}
