% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLASheader.r
\name{readLASheader}
\alias{readLASheader}
\title{Read LiDAR header}
\usage{
readLASheader (LASfile)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}
}
\value{
Returns a list of the information stored in the LAS file header.
}
\description{
This function reads and returns values associated with the LAS file format header. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{

}
\author{
Original Script: rLiDAR package by Michael Sumner and Carlos Alberto Silva. https://cran.r-project.org/package=rLiDAR Modified by Nicholas Kruskamp to read and output the header information.
}
