% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLASvlr.r
\name{readLASvlr}
\alias{readLASvlr}
\title{Read LiDAR vlr}
\usage{
readLASvlr (LASfile)
}
\arguments{
\item{LASfile}{A standard LAS data file (ASPRS)}
}
\value{
Returns a list of lists with vlr information
}
\description{
This function reads and returns values associated with the LAS file Variable Length Records (vlr). The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS). This tool does not currently support Extended vlr.
}
\examples{

}
\author{
Original Script: rLiDAR package by Michael Sumner and Carlos Alberto Silva. https://cran.r-project.org/package=rLiDAR Modified by Nicholas Kruskamp to read vlr.
}
