% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertDistRatio.r
\name{vertDistRatio}
\alias{vertDistRatio}
\title{Calculate Vertical Distribution Ratios}
\usage{
vertDistRatio(x, resolution = 30, pointClasses = c(3,4,5))
}
\arguments{
\item{x}{Spatial Points Data Frame containing X, Y, Z coordinates and Classification data.}

\item{resolution}{Number specifying the grid cell resolution of the output raster.}

\item{pointClasses}{List of numbers specifying the point classifications to subset the lidar data.}
}
\value{
returns a raster stack where each layer is a ratio.
}
\description{
This function calculates two vertical distribution ratios for vegetation analysis from lidar data. The ratios are calulated as the 98%ile - 50%ile / 98%ile, and 100%ile - 50%ile / 100%ile.
}
\examples{

}
\author{
Nicholas Kruskamp
}
