% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLidarData.R
\name{readLidarData}
\alias{readLidarData}
\title{Read LiDAR point data}
\usage{
readLidarData (LASfile, inputCRS, maxHAG = 70, returnAll=FALSE)
}
\arguments{
\item{LASfile}{A standard LAS or LAZ data file (ASPRS)}

\item{proj4string}{to be interpreted as a CRS by SP pacakge.}

\item{Logical.}{Default FALSE. By default, only returns Classification 0, 1, & 2.}
}
\value{
Returns a Spatial Points class of the point information stored in the LAS file.
}
\description{
This function reads and returns values associated with the LAS file format. The LAS file is a public file format for the interchange of LiDAR 3-dimensional point cloud data (American Society of Photogrammetry and Remote Sensing - ASPRS)
}
\examples{

}
\author{
Nicholas Kruskamp
}
